autorizacão

+++ app/Http/Controllers/ProjectController.php
@@ -4,6 +4,7 @@ namespace Projeto\Http\Controllers;
 
 
 use Illuminate\Http\Request;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 use Projeto\Repositories\ProjectRepository;
 
 use Projeto\Services\ProjectService;
@@ -32,7 +33,8 @@ class ProjectController extends Controller
     }
     public function index()
     {
-       return  $this->repository->all();
+
+       return  $this->repository->findWhere(['owner_id'=>\Authorizer::getResourceOwnerId()]);
 
     }
 
@@ -57,6 +59,10 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+        if($this->checkProjectPermission($id)==false)
+        {
+            return ['error'=>'Access Forbidden'];
+        }
         return $this->repository->find($id);
     }
 
@@ -71,6 +77,10 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if($this->checkProjectPermission($id)==false)
+        {
+            return ['error'=>'Access Forbidden'];
+        }
         return $this->service->update($request->all(),$id);
     }
 
@@ -82,7 +92,30 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if($this->checkProjectPermission($id)==false)
+        {
+            return ['error'=>'Access Forbidden'];
+        }
        return $this->repository->delete($id);
 
     }
+    private function checkProjectOwner($projectId)
+    {
+        $userId= \Authorizer::getResourceOwnerId();
+        return $this->repository->isOwner($projectId,$userId);
+    }
+    private function checkProjectMember($projectId)
+    {
+        $userId= \Authorizer::getResourceOwnerId();
+        return $this->repository->hasMember($projectId,$userId);
+    }
+    private function checkProjectPermission($projectId)
+    {
+        if($this->checkProjectOwner($projectId)or $this->checkProjectMember($projectId))
+        {
+            return true;
+        }
+        return false;
+    }
+
 }
